+++ amd/src/editor.js
@@ -34,6 +34,81 @@ define(['jquery'], function($) {
         });
     };
 
+    var createFullScreenButtons = function() {
+        require(['core/str'], function(str) {
+            var enterFullScreenText = str.get_string('editorenterfullscreen', 'onlyofficeeditor');
+            var exitFullScreenText = str.get_string('editorexitfullscreen', 'onlyofficeeditor');
+            var navButton = $('nav .nav-link.btn')[0];
+            var editorContainer = $('.onlyofficeeditor-container')[0];
+
+            $.when(enterFullScreenText).done(function(localized) {
+                enterFullScreenText = localized;
+                var enterButton = document.createElement('button');
+                var enterIcon = document.createElement('i');
+                enterIcon.className = 'icon fa fa-expand fa-fw';
+                enterButton.appendChild(enterIcon);
+                enterButton.className = 'onlyofficeeditor-editor-fs-button';
+                enterButton.id = 'onlyofficeeditor-enter-fs-button';
+                enterButton.innerHTML += enterFullScreenText;
+
+                enterButton.onclick = function() {
+                    $('header').hide();
+                    $('footer').hide();
+                    if (navButton.getAttribute('aria-expanded') === 'true') {
+                        $('nav .nav-link.btn')[0].click();
+                    }
+                    editorContainer.style.cssText = 'position: absolute; left: 0; right: 0; top: 0; ' +
+                        'padding: 0 16px 0 16px; z-index: 100;';
+                    editorContainer.children[0].style.height = '93.5vh';
+                    $('#onlyofficeeditor-enter-fs-button').hide();
+                    $('#onlyofficeeditor-exit-fs-button').show();
+                };
+                $("#region-main-settings-menu .menubar")[0].prepend(enterButton);
+            });
+            $.when(exitFullScreenText).done(function(localized) {
+                exitFullScreenText = localized;
+                var exitButton = document.createElement('button');
+                var exitIcon = document.createElement('i');
+                exitIcon.className = 'icon fa fa-compress fa-fw';
+                exitButton.appendChild(exitIcon);
+                exitButton.innerHTML += exitFullScreenText;
+                exitButton.className = 'onlyofficeeditor-editor-fs-button';
+                exitButton.id = 'onlyofficeeditor-exit-fs-button';
+
+                exitButton.onclick = function() {
+                    editorContainer.style.cssText = '';
+                    editorContainer.children[0].style.height = '95vh';
+                    $('header').show();
+                    $('footer').show();
+                    $('#onlyofficeeditor-enter-fs-button').show();
+                    $('#onlyofficeeditor-exit-fs-button').hide();
+                };
+                $('.usernav .nav-item')[0].prepend(exitButton);
+                $('#onlyofficeeditor-exit-fs-button').hide();
+            });
+        });
+    };
+
+    $.urlParam = function(name) {
+        var results = new RegExp('[\\?&]' + name + '=([^&#]*)').exec(window.location.href);
+        if (results === null) {
+            return null;
+        }
+        return decodeURI(results[1]) || 0;
+    };
+
+    var replaceActionLink = function(href, linkParam) {
+        var link;
+        var actionIndex = href.indexOf("&actionType=");
+        if (actionIndex != -1) {
+            link = href.substring(0, actionIndex) + "&actionType=" + encodeURIComponent(linkParam.type) +
+                "&actionData=" + encodeURIComponent(linkParam.data);
+        } else {
+            link = href + "&actionType=" + encodeURIComponent(linkParam.type) + "&actionData=" + encodeURIComponent(linkParam.data);
+        }
+        return link;
+    };
+
     var saveAsModal = null;
 
     var displaySaveAsModal = function(saveAsData, cmid, courseid) {
@@ -61,10 +136,13 @@ define(['jquery'], function($) {
                 displayNotification('docserverunreachable', 'error');
                 return;
             }
+            createFullScreenButtons();
             var ajaxUrl = M.cfg.wwwroot + '/mod/onlyofficeeditor/dsconfig.php';
             $.getJSON(ajaxUrl, {
                 courseid: courseid,
-                cmid: cmid
+                cmid: cmid,
+                actionType: $.urlParam('actionType'),
+                actionData: $.urlParam('actionData')
             }).done(function(data) {
                 var docEditor = null;
                 var config = data.config;
@@ -88,6 +166,33 @@ define(['jquery'], function($) {
                         docEditor.showMessage(message);
                     }
                 };
+
+                var onMakeActionLink = function(event) {
+                    var actionData = event.data.action;
+                    docEditor.setActionLink(replaceActionLink(location.href, actionData));
+                };
+
+                var onRequestSendNotify = function(event) {
+                    var comment = event.data.message;
+                    var emails = event.data.emails;
+                    var replacedActionLink = replaceActionLink(location.href, event.data.actionLink.action);
+
+                    var mentionData = {
+                        comment: comment,
+                        emails: emails,
+                        link: replacedActionLink,
+                        courseid: courseid
+                    };
+
+                    $.ajax(M.cfg.wwwroot + '/mod/onlyofficeeditor/onlyofficeeditorapi.php?apiType=mention&cmid=' + cmid, {
+                        type: 'POST',
+                        dataType: 'json',
+                        data: mentionData
+                    }).fail(() => {
+                        displayNotification('onmentionerror', 'error');
+                    });
+                };
+
                 const onAppReady = () => {
                     innerAlert("Document editor ready");
                 };
@@ -99,7 +204,8 @@ define(['jquery'], function($) {
                 };
                 config.events = {
                     'onAppReady': onAppReady,
-                    'onError': onError
+                    'onError': onError,
+                    'onMakeActionLink': onMakeActionLink
                 };
 
                 var onRequestSaveAs = function(event) {
@@ -114,6 +220,17 @@ define(['jquery'], function($) {
 
                 if (canAddInstance) {
                     config.events.onRequestSaveAs = onRequestSaveAs;
+                }
+
+                var usersToMention = data.userstomention;
+
+                var onRequestUsers = function() {
+                    docEditor.setUsers({'users': usersToMention});
+                };
+
+                if (usersToMention !== null) {
+                    config.events.onRequestUsers = onRequestUsers;
+                    config.events.onRequestSendNotify = onRequestSendNotify;
                 }
 
                 if ((config.document.fileType === "docxf" || config.document.fileType === "oform")