+++ classes/editor.php
@@ -135,6 +135,7 @@ class editor {
         // Editorconfig.
         $editorconfig = [];
         $pathnamehash = $crypt->get_hash(['userid' => $USER->id, 'pathnamehash' => $file->get_pathnamehash(), 'cm' => $this->cm]);
+        $editorconfig['actionLink'] = null;
         $editorconfig['callbackUrl'] = $CFG->wwwroot . '/mod/onlyofficeeditor/callback.php?doc=' . $pathnamehash;
         $editorconfig['lang'] = $USER->lang;