+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## Added
+- add protect option for editor
+- add ability to open editor in fullscreen mode
+- add mentions in comments in editor
 - "save as" in editor
 - add ability to create activity with empty file
 - add favicon on editor page
@@ -8,6 +11,7 @@
 
 ## Changed
 - set editor interface language
+- fixes building backup and restore
 
 ## 2.1.0
 ## Added